abstraction for ZeroPadding
